Улучшение стиля программы и готовый Пункт № 18.
