+++ Life/Life.csproj
@@ -50,9 +50,11 @@
             <Generator>MSBuild:Compile</Generator>
             <SubType>Designer</SubType>
         </ApplicationDefinition>
+        <Compile Include="src\Evolution.cs" />
         <Compile Include="src\Game.cs" />
         <Compile Include="src\InitGame.cs" />
         <Compile Include="src\Picture.cs" />
+        <Compile Include="src\SLCondition.cs" />
         <Page Include="MainWindow.xaml">
             <Generator>MSBuild:Compile</Generator>
             <SubType>Designer</SubType>
