+++ Life/Readme.md
@@ -14,14 +14,14 @@
 * Пункт №12
 * Пункт №13
 * Пункт №14
+* Пункт №18
 
 ## Максимально возможное количество баллов
 
-32 балла
+34   балла
 
 ## Надо бы сделать
 * Пункт №11 вторую половину - примерно 1 балл
-* Пункт №18 Автоматическое проигрывание автоматически останавливается, если на очередном ходу позиция не поменялась — 2 балла
 * Пункт №17 (по возможности) - 12 баллов
 * Перекинуть все Эволвы в другой файл
 * gridMouseDown переделать под функции
